fix(numeric): read01 round 271 — NUMERIC's display scale is a u16
At u8 a decimal with more than 255 places could not be represented at
all, and the failure was not a clean one: a plain 256-place decimal
reached a converter whose expect("lexer-validated decimal") aborted the
query with an internal error, on SQL PG accepts. The literal path had
its own escape hatch, quietly handing such a value to the float lexer,
so pg_typeof(1e-256) answered double precision where PG says numeric.
Round 270 hit the second half of this and recorded it; this closes both.

The same 255 ceiling had been written FOUR separate times — in the
lexer's decimal scanner, the engine's numeric parser, the big-decimal
parser, and the wire's decimal_str_to_scaled. Each one independently
decided a longer decimal was somebody else's problem.

Widened through the stack: the value's scale, BigNumeric's, the
declared type's, the AST literal's, NUMERIC[]'s per-element scale, and
both wire paths. On disk nothing already written moves — a scale that
still fits a byte keeps its existing shape, and the scales that could
not previously EXIST take new form bytes (numeric form 4, big-numeric
tag 34, array element flag 2) and a new DataType tag 69. No migration,
no version fence.

Found while widening: DataType::Real wrote catalog tag 66, which PgLsn
also writes and which the reader maps to PgLsn, so a persisted REAL
column came back as pg_lsn. The collision was unreachable until round
269 made REAL a column type. Real now writes tag 68.

Three arithmetic paths assumed the old ceiling and became reachable:

  * pow10_i128 (two of them, in numeric.rs and binop.rs) multiplied
    unchecked and panicked. numeric.rs's now saturates like its
    pow10_sat sibling; binop's rescale returns None so the caller drops
    onto the BigNumeric path it already has.
  * round/trunc with a target scale past what i128 can carry fell
    through to the f64 tail and handed back double precision, as the
    negative-scale case used to before round 98/99 fixed that half.
  * numeric -> float built its divisor by multiplying by ten in a loop.
    That accumulated rounding error (1e-300 came out
    9.999999999999999e-301) and ran to infinity once the scale could
    exceed 308, which then looked like an underflow and errored. All
    four copies now parse the decimal text.

That last one also fixed mixed float/numeric arithmetic:
1e-300::float8 - 1e-300 answered 1.6578092e-316 instead of 0 because
the conversion, not the arithmetic, was wrong.

Also: scale() had no arm for the arbitrary-precision form, which only
became reachable now that round(1e-256, 300) produces one.

Two ceilings are PG's and are now honoured rather than clipped by u8:
the division and sqrt display scales cap at PG's 1000, not 255.

Recorded residual: NUMERIC(p,s) still caps p at 38 (i128), where PG
allows 1000. That is the mantissa's width rather than the scale's, so
it is its own question.

Gate: 6454 passed / 0 failed, clippy back to the 5 pre-existing
warnings.

+++ crates/spg-engine/src/aggregate.rs
@@ -368,7 +368,7 @@ pub(crate) struct AggState {
     /// the running sum as an integral mantissa at `sum_num_scale`
     /// (aligned on the max input scale, so it stays exact — no f64).
     sum_num_scaled: i128,
-    sum_num_scale: u8,
+    sum_num_scale: u16,
     /// v7.39 (read01 numeric.c) — bignum spill for a sum that leaves i128;
     /// see `SumBig`. The i128 lane freezes at zero once this engages.
     sum_big: SumBig,
@@ -1074,7 +1074,7 @@ struct FusedAcc {
     float_not_real: bool,
     num_scaled: i128,
     num_kind: spg_storage::NumericKind,
-    num_scale: u8,
+    num_scale: u16,
     /// v7.39 (read01 numeric.c) — bignum spill; see `SumBig`.
     num_big: SumBig,
     use_numeric: bool,
@@ -1235,10 +1235,10 @@ type SumBig = Option<alloc::boxed::Box<spg_storage::bignum::BigNumeric>>;
 /// Add an exact NUMERIC (mantissa × 10^-scale) into the sum tri-state.
 fn sum_add_exact(
     scaled: &mut i128,
-    scale: &mut u8,
+    scale: &mut u16,
     big: &mut SumBig,
     add_scaled: i128,
-    add_scale: u8,
+    add_scale: u16,
 ) {
     use spg_storage::bignum::BigNumeric;
     if let Some(b) = big {
@@ -1264,7 +1264,7 @@ fn sum_add_exact(
 /// Add a BigNumeric input into the sum tri-state (promotes immediately).
 fn sum_add_bignum(
     scaled: &mut i128,
-    scale: &mut u8,
+    scale: &mut u16,
     big: &mut SumBig,
     b_in: &spg_storage::bignum::BigNumeric,
 ) {
@@ -1973,7 +1973,7 @@ fn accumulate_groups(
         // nothing for this.
         let mut num_scaled: i128 = 0;
         let mut num_kind = spg_storage::NumericKind::Finite;
-        let mut num_scale: u8 = 0;
+        let mut num_scale: u16 = 0;
         let mut num_big: SumBig = None;
         let mut use_numeric = false;
         // Interval accumulator (sum/avg over an interval column).
@@ -4791,7 +4791,7 @@ fn values_to_array(picked: &[Value<'_>]) -> Value<'static> {
 /// percentile_cont). `scaled × 10^-scale`; `10^scale` fits in i128 for the
 /// NUMERIC scale range, so no `f64::powi` (unavailable under no_std) is needed.
 #[allow(clippy::cast_precision_loss)]
-fn numeric_to_f64(scaled: i128, scale: u8) -> f64 {
+fn numeric_to_f64(scaled: i128, scale: u16) -> f64 {
     (scaled as f64) / (10i128.pow(u32::from(scale)) as f64)
 }
 
@@ -6012,7 +6012,7 @@ mod value_cmp_mixed_numeric_tests {
     use core::cmp::Ordering;
     use spg_storage::Value;
 
-    fn num(scaled: i128, scale: u8) -> Value<'static> {
+    fn num(scaled: i128, scale: u16) -> Value<'static> {
         Value::Numeric {
             scaled,
             scale,
